+++ examples/live/binance/binance_spot_demo_exec_tester.py
@@ -22,77 +22,52 @@ from nautilus_trader.adapters.binance import BinanceDataClientConfig
 from nautilus_trader.adapters.binance import BinanceExecClientConfig
 from nautilus_trader.adapters.binance import BinanceLiveDataClientFactory
 from nautilus_trader.adapters.binance import BinanceLiveExecClientFactory
-from nautilus_trader.cache.config import CacheConfig
+from nautilus_trader.adapters.binance.common.enums import BinanceEnvironment
 from nautilus_trader.config import InstrumentProviderConfig
 from nautilus_trader.config import LiveExecEngineConfig
 from nautilus_trader.config import LoggingConfig
 from nautilus_trader.config import TradingNodeConfig
-from nautilus_trader.examples.strategies.ema_cross import EMACross
-from nautilus_trader.examples.strategies.ema_cross import EMACrossConfig
 from nautilus_trader.live.node import TradingNode
-from nautilus_trader.model.data import BarType
 from nautilus_trader.model.identifiers import InstrumentId
 from nautilus_trader.model.identifiers import TraderId
+from nautilus_trader.test_kit.strategies.tester_exec import ExecTester
+from nautilus_trader.test_kit.strategies.tester_exec import ExecTesterConfig
 
 
 # *** THIS IS A TEST STRATEGY WITH NO ALPHA ADVANTAGE WHATSOEVER. ***
 # *** IT IS NOT INTENDED TO BE USED TO TRADE LIVE WITH REAL MONEY. ***
 
 # Strategy config params
-symbol = "ETHUSDT-PERP"
+symbol = "ETHUSDT"
 instrument_id = InstrumentId.from_str(f"{symbol}.{BINANCE}")
-order_qty = Decimal("0.02")
+order_qty = Decimal("0.01")
 
 # Configure the trading node
 config_node = TradingNodeConfig(
     trader_id=TraderId("TESTER-001"),
     logging=LoggingConfig(log_level="INFO"),
     exec_engine=LiveExecEngineConfig(
-        # debug=True,
         reconciliation=True,
         reconciliation_lookback_mins=1440,
-        # snapshot_orders=True,
-        # snapshot_positions=True,
-        # snapshot_positions_interval_secs=5.0,
+        open_check_interval_secs=0.0,  # TODO: configure appropriate interval
     ),
-    cache=CacheConfig(
-        # database=DatabaseConfig(timeout=2),
-        timestamps_as_iso8601=True,
-        flush_on_start=False,
-    ),
-    # message_bus=MessageBusConfig(
-    #     database=DatabaseConfig(timeout=2),
-    #     timestamps_as_iso8601=True,
-    #     use_instance_id=False,
-    #     # types_filter=[QuoteTick],
-    #     stream_per_topic=False,
-    #     external_streams=["bybit"],
-    #     autotrim_mins=30,
-    # ),
     data_clients={
         BINANCE: BinanceDataClientConfig(
-            api_key=None,  # 'BINANCE_API_KEY' env var
-            api_secret=None,  # 'BINANCE_API_SECRET' env var
-            account_type=BinanceAccountType.USDT_FUTURES,
-            base_url_http=None,  # Override with custom endpoint
-            base_url_ws=None,  # Override with custom endpoint
-            us=False,  # If client is for Binance US
-            testnet=True,  # If client uses the testnet
+            api_key=None,  # 'BINANCE_DEMO_API_KEY' env var
+            api_secret=None,  # 'BINANCE_DEMO_API_SECRET' env var
+            account_type=BinanceAccountType.SPOT,
+            environment=BinanceEnvironment.DEMO,
             instrument_provider=InstrumentProviderConfig(load_all=True),
         ),
     },
     exec_clients={
         BINANCE: BinanceExecClientConfig(
-            api_key=None,  # 'BINANCE_API_KEY' env var
-            api_secret=None,  # 'BINANCE_API_SECRET' env var
-            account_type=BinanceAccountType.USDT_FUTURES,
-            base_url_http=None,  # Override with custom endpoint
-            base_url_ws=None,  # Override with custom endpoint
-            us=False,  # If client is for Binance US
-            testnet=True,  # If client uses the testnet
+            api_key=None,  # 'BINANCE_DEMO_API_KEY' env var
+            api_secret=None,  # 'BINANCE_DEMO_API_SECRET' env var
+            account_type=BinanceAccountType.SPOT,
+            environment=BinanceEnvironment.DEMO,
             instrument_provider=InstrumentProviderConfig(load_all=True),
             max_retries=3,
-            use_position_ids=False,
         ),
     },
     timeout_connection=30.0,
@@ -106,21 +81,20 @@ config_node = TradingNodeConfig(
 node = TradingNode(config=config_node)
 
 # Configure your strategy
-strat_config = EMACrossConfig(
+config_strat = ExecTesterConfig(
     instrument_id=instrument_id,
     external_order_claims=[instrument_id],
-    bar_type=BarType.from_str(f"{instrument_id}-1-MINUTE-LAST-EXTERNAL"),
-    fast_ema_period=10,
-    slow_ema_period=20,
-    trade_size=order_qty,
-    order_id_tag="001",
-    oms_type="HEDGING",
-    subscribe_trade_ticks=True,
-    subscribe_quote_ticks=True,
+    order_qty=order_qty,
+    open_position_on_start_qty=order_qty,
+    enable_limit_buys=True,
+    enable_limit_sells=True,
+    use_post_only=True,
+    # tob_offset_ticks=0,
+    log_data=False,
 )
 
 # Instantiate your strategy
-strategy = EMACross(config=strat_config)
+strategy = ExecTester(config=config_strat)
 
 # Add your strategies and modules
 node.trader.add_strategy(strategy)